Fixed a bug when loading the config on log-in.
GetCharacterID returns nil when the PLAYER_ENTERED_GAME event is called
which was causing an error with loading the configuration. I set a
timeout to try again for as long as it's returning nil. In my testing it
only took one iteration, so I think this is a race condition.