plod: Partial Trade Assist integration.
I implemented annoyingly convoluted Trade Assist integration. It
basically works, but in the end the API calls don't notify the user what
the profit is on their sales. I think I need to track that information
also, but plod's already starting to look like a hodgepodge.

I'm going to rearchitect it on some level. I want to commit this because
it's pretty close and represents a lot of work in figuring out the VO's
Lua API.