Support for array routes and function handlers.
You can now pass in an array of arrays to the contructor, if you prefer not to
have to load an ini file.

The router now checks for the '::' delimeter. If it's not there it assumes that
the handler is a function. I'm going to use this in some non-OOP code.