Support for multiple apps in the same database.
I'm using dbdo in a project that has a web frontend to dbdo and I'd like the
ability to run and unrun fixtures without affecting the schema migrations of
the main application. All the functions have an additional argument for the
application name. This works as a suffix to the database table that tracks the
migrations.