Fixed the handling of initial windows and panes.
The environment variables weren't being set in the first window, first pane.
This was due to the fact that it already exists when the set-environment
commands are called. Now, after the environment set up, we'll create a new
window, delete the initial one and move the new one into place.

This also brought to light a mis-handling of the base-index and base-pane-index
properties. We'll now determine these values before window/pane manipulation.