Added TENPER_VIRTUALENV to tenper sessions.
This is the path to the virtualenv activate script. I'm not going to document
this at the moment because I'm mulling stuff over.

I use this because we can't automatically source this script for every new
pane. It might be nice instead to augment the tenper commands with stuff like:

  tenper source-virtualenv
  tenper new-window