Initial commit.
I reset the repository because I had a bunch of temp commits. Since I'm making
this a literate programming, the history of the repository should be less
important. Usually we look at the history to understand why a change was made,
but a literate program should explain all this in the document itself.