Compile files in the shell process.
I wasn't getting errors when LFE files failed to compile so I switched to using
the `.erlang` file to start a process that would compile files. Since it's
running in the shell errors are printed in the shell.

As part of making the system a little nicer, I also no longer support a watcher
*and* a shell. It's been years since I've developed without one open.